:art: refactor: FeedCardList hook 파일 생성 및 코드 리팩토링
