+++ src/App.jsx
@@ -1,4 +1,6 @@
+import FeedCardList from './components/FeedCardList/FeedCardList';
+
 function App() {
-  return <></>;
+  return <FeedCardList subjectId='6742' />;
 }
 export default App;
