--- src/pages/MainPage.js
@@ -1,1 +1,0 @@
-main;
